+++ README.md
@@ -9,12 +9,15 @@
 * [텍스트와 이미지로 살펴보는 온라인 코드 리뷰 과정](https://github.com/next-step/nextstep-docs/tree/master/codereview)
 
 ## 구현목록
-- [ ] 경주 조건을 입력받는다
-- [ ] 게임을 만든다
-- [ ] 게임을 시작한다
-- [ ] 주어진 횟수만큼의 이동을 시도한다
-- [ ] 전진 조건을 판단한다
-- [ ] 전진 조건을 만족하면 진한다
-- [ ] 멈춘다
-- [ ] 경주 결과를 출력한다
+- [x] 경주 게임 진행을 위한 조건을 입력받는다
+- [x] 이름을 입력 받는다 
+- [x] 이름의 길이를 제한한다
+- [x] 이름을 쉼표로 구분한다
+- [x] 게임을 만든다
+- [x] 게임을 시작한다
+- [x] 주어진 횟수만큼의 이동을 시도한다
+- [x] 전진 조건을 판단한다
+- [x] 전진 조건을 만족하면 진한다
+- [x] 멈춘다
+- [x] 경주 결과를 출력한다
