--- src/main/java/racing_car_with_winner/README.md
@@ -1,12 +1,0 @@
-# 4단계 - 자동차 경주(우승자) 
-
-
-## 기능목록  
-- 경주 진행을 위한 조건을 입력 받는다
-  - 이름을 입력받는다 
-  - 이름은 쉼표로 구분된다 
-  - 이름의 길이를 제한한다 
-  - 시도할 회수를 입력 받는다  
-- 자동차는 전진하거나 멈출 수 있다 
-- 경주 결과를 출력한다
-- 우승자를 선정한다
