+++ src/main/java/racing_car_with_winner/model/Car.java
@@ -1,8 +1,6 @@
-package racing_car_with_winner.model.car;
+package racing_car_with_winner.model;
 
 import java.util.Objects;
-
-import racing_car_with_winner.model.MovableStrategy;
 
 public class Car {
     private final CarName carName;
