+++ src/main/java/racing_car_with_winner/model/Cars.java
@@ -1,11 +1,9 @@
-package racing_car_with_winner.model.car;
+package racing_car_with_winner.model;
 
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-
-import racing_car_with_winner.model.MovableStrategy;
 
 public class Cars {
     private final List<Car> cars;
