+++ src/test/java/racing_car_with_winner/model/CarFactoryTest.java
@@ -4,6 +4,8 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
+import racing_car_with_winner.view.CarNameInput;
+
 class CarFactoryTest {
 
     @ParameterizedTest
