+++ src/test/java/racing_car_with_winner/model/CarNameInputTest.java
@@ -3,6 +3,8 @@ package racing_car_with_winner.model;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import racing_car_with_winner.view.CarNameInput;
+
 class CarNameInputTest {
 
     @Test
