+++ src/test/java/racing_car_with_winner/model/CarsTest.java
@@ -6,10 +6,7 @@ import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import racing_car_with_winner.model.car.Car;
-import racing_car_with_winner.model.car.CarName;
-import racing_car_with_winner.model.car.Cars;
-import racing_car_with_winner.model.car.MoveResult;
+import racing_car_with_winner.view.CarNameInput;
 
 public class CarsTest {
