+++ src/test/java/racing_car_with_winner/model/MoveResultTest.java
@@ -3,8 +3,6 @@ package racing_car_with_winner.model;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import racing_car_with_winner.model.car.MoveResult;
-
 class MoveResultTest {
     @Test
     void compare() {
